Adding module SwitchFilter, that disables PacketIn and LLDP for some switches
This commit adds a Simple Floodlight Module that drops all PacketIn messages
from switches with DPID matching given values. It also disables LLDP to those
switches as soon as they start a connection to Floodlight.

To add a drop filter, just call filterMatchSet.add(String s) in this module's
init function. The String s is a regex pattern that will be matched to each
switch DPID, with the module disabling PacketIn and LLDP case the match
succeed.